etl_pipeline/requirements.txt
dagster 
dagster-webserver 
dagit
dagster-postgres 
SQLAlchemy
pandas 
minio 
mysql-connector-python
psycopg2-binary
python-dotenv
